--- authentication/models.py
@@ -1,40 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import AbstractUser, PermissionsMixin, BaseUserManager
-
-
-class CustomUserManager(BaseUserManager):
-
-    def create_user(self, email, password, **extra_fields):
-        if not email:
-            raise ValueError('email Field is required')
-
-        email = self.normalize_email(email)
-        user = self.model(email=email, **extra_fields)
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password, **extra_fields):
-        extra_fields.setdefault('is_staff', True)
-        extra_fields.setdefault('is_active', True)
-        extra_fields.setdefault('is_superuser', True)
-
-        if extra_fields.get('is_staff') is not True:
-            raise ValueError('superuser must have is_staff=True')
-
-        if extra_fields.get('is_active') is not True:
-            raise ValueError('superuser must have is_active=True')
-
-        return self._create_user(email, password, **extra_fields)
-
-
-class CustomUser(AbstractUser, PermissionsMixin):
-    email = models.EmailField(max_length=100, unique=True)
-
-    USERNAME_FIELD = 'email'
-    REQUIRED_FIELDS = ['username']
-
-    objects = CustomUserManager()
-
-    def __str__(self):
-        return self.email
