+++ requirements.txt
@@ -5,5 +5,6 @@ python-decouple==3.6
 dj-database-url==0.5.0
 psycopg2-binary==2.9.3
 djangorestframework-simplejwt==5.0.0
-
-
+gunicorn==20.1.0
+whitenoise==6.2.0
+django-cors-headers==3.13.0
