+++ runtime.txt
@@ -1,0 +1,1 @@
+Python 3.10.5
